+++ test/Amadevus.RecordGenerator.TestsBase/StructContainer.cs
@@ -1,0 +1,8 @@
+﻿namespace Amadevus.RecordGenerator.TestsBase
+{
+    [Record(Features.Default | Features.Equality)]
+    public readonly partial struct StructContainer 
+    {
+        public int Id { get; }
+    }
+}